% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrgl.R
\name{silicate-mesh3d-methods}
\alias{silicate-mesh3d-methods}
\alias{sc_vertex.mesh3d}
\alias{sc_coord.mesh3d}
\title{Silicate methods}
\usage{
\method{sc_vertex}{mesh3d}(x, ..., prefer_triangles = TRUE)

\method{sc_coord}{mesh3d}(x, ..., prefer_triangles = TRUE)
}
\description{
Silicate methods
}
